requirements-mcp.txt
# MCP (Model Context Protocol) + MySQL
fastmcp
pymysql
python-dotenv
pyyaml
